Fix long delay sending mail when using pinned SMTP server cert. Bug 775730
Since glib-networking actually integrates with the p11f-kit
infrastructure for the default GNU TLS backend, if we just set the GCR
peer name correctly when pinning a certificate, it'll automatically check
for pinned certs for us.

This lets us remove the code currently checking for a pinned cert in the
main loop that was causing a re-connect when encountering a pinned cert,
but at the cost of no longer being able to unpin certs from the app any
more. This is a tradeoff I'm happy with.

* src/client/application/geary-controller.vala (GearyController): Use new
  GcrService class for managing cert pinning. Remove check for pinned
  cert on TLS error and (possibly sketchy anyway) code unpin existing
  certs.

* src/client/application/gcr-service.vala: New class for managing GCR
  code in the one place.

* src/client/application/geary-args.vala (Args): Remobe the revoke-certs
  arg and global var.

* src/CMakeLists.txt: Add new source file.